Module count_vowels::count_vowels

A function that will convert a request into a response that can be used to submit a request to the server. This function counts the number of vowels in the text and it reports a sum of each vowel found.

Notes

This could be written completly in JavaScript but we’re testing interacting with the server and server side code. So we will use little or no JavaScript when possible.

Structs

XFrameOptions
Avoid Clickjacking attacks, by ensuring that their content is not embedded into other sites.

XXssProtection
Force/Disable XSS protection

Functions

get_response
Accept a HTTP request and return the appropriate HTTP response.